chore(tools): one browser finder instead of two drifting copies
find_chrome() existed in check_dashboard_layout.py and make_screenshots.py, each with
its own browser list and mac path, and with different contracts -- one returned None,
the other raised. The drift that invites is specific: someone adds a browser name to
one copy, and the two tools disagree about which browser they find.

One definition now, in build_web -- the module every page-rendering tool already
imports, so this adds no coupling that did not exist. The softer contract (None) is
the shared one; make_screenshots raises at its own call site with its own message, so
both command-line behaviours are unchanged.

The one finding of the 2026-07-29 cleanup audit. The rest of that audit found nothing
to change: zero own-code warnings on cold builds, no dead code, no unused defines, no
stdlib-replaceable hand-rolled logic, and every dependency pin already at its upstream
latest. The remaining deprecation warnings live in eModbus and the framework, and
fixing those means forking a dependency this project has decided to keep.
