+++ tools/check_dashboard_layout.py
@@ -26,7 +26,6 @@
 
 import pathlib
 import re
-import shutil
 import subprocess
 import sys
 import tempfile
@@ -649,20 +648,6 @@ const tick=setInterval(async()=>{
 """
 
 
-def find_chrome() -> str | None:
-    for name in (
-        "google-chrome",
-        "google-chrome-stable",
-        "chromium",
-        "chromium-browser",
-    ):
-        found = shutil.which(name)
-        if found:
-            return found
-    mac = "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome"
-    return mac if pathlib.Path(mac).exists() else None
-
-
 def build_page(stripped: str, stub: str, battery: str, extra_js: str) -> str:
     """The served page with the stub attached, and the assertions after it.
 
@@ -717,7 +702,7 @@ def main() -> int:
     stripped = build_web.served_page()
     stub = (ROOT / "tools" / "demo_fleet.js").read_text(encoding="utf-8")
 
-    chrome = find_chrome()
+    chrome = build_web.find_chrome()
     if chrome is None:
         print("dashboard layout: FAIL (no Chrome/Chromium on PATH to render with)")
         return 1
